uncompressed: pulled MixedInterleaveDecoder into its own file
