+++ libheif/codecs/uncompressed/unc_codec.cc
@@ -39,6 +39,7 @@
 #include "decoder_abstract.h"
 #include "decoder_component_interleave.h"
 #include "decoder_pixel_interleave.h"
+#include "decoder_mixed_interleave.h"
 
 
 bool isKnownUncompressedFrameConfigurationBoxProfile(const std::shared_ptr<const Box_uncC>& uncC)
@@ -395,114 +396,6 @@ bool map_uncompressed_component_to_channel(const std::shared_ptr<const Box_cmpd>
 }
 
 
-class MixedInterleaveDecoder : public AbstractDecoder
-{
-public:
-  MixedInterleaveDecoder(uint32_t width, uint32_t height, std::shared_ptr<Box_cmpd> cmpd, std::shared_ptr<Box_uncC> uncC) :
-      AbstractDecoder(width, height, std::move(cmpd), std::move(uncC)) {}
-
-  Error decode_tile(const HeifContext* context,
-                    heif_item_id image_id,
-                    std::shared_ptr<HeifPixelImage>& img,
-                    uint32_t out_x0, uint32_t out_y0,
-                    uint32_t image_width, uint32_t image_height,
-                    uint32_t tile_x, uint32_t tile_y) override
-  {
-    if (m_tile_width == 0) {
-      return {heif_error_Decoder_plugin_error, heif_suberror_Unspecified, "Internal error: MixedInterleaveDecoder tile_width=0"};
-    }
-
-    // --- compute which file range we need to read for the tile
-
-    uint64_t tile_size = 0;
-
-    for (ChannelListEntry& entry : channelList) {
-      if (entry.channel == heif_channel_Cb || entry.channel == heif_channel_Cr) {
-        uint32_t bits_per_row = entry.bits_per_component_sample * entry.tile_width;
-        bits_per_row = (bits_per_row + 7) & ~7U; // align to byte boundary
-
-        tile_size += bits_per_row / 8 * entry.tile_height;
-      }
-      else {
-        uint32_t bits_per_component = entry.bits_per_component_sample;
-        if (entry.component_alignment > 0) {
-          uint32_t bytes_per_component = (bits_per_component + 7) / 8;
-          bytes_per_component += nAlignmentSkipBytes(entry.component_alignment, bytes_per_component);
-          bits_per_component = bytes_per_component * 8;
-        }
-
-        uint32_t bits_per_row = bits_per_component * entry.tile_width;
-        bits_per_row = (bits_per_row + 7) & ~7U; // align to byte boundary
-
-        tile_size += bits_per_row / 8 * entry.tile_height;
-      }
-    }
-
-
-    if (m_uncC->get_tile_align_size() != 0) {
-      tile_size += nAlignmentSkipBytes(m_uncC->get_tile_align_size(), tile_size);
-    }
-
-    assert(m_tile_width > 0);
-    uint32_t tileIdx = tile_x + tile_y * (image_width / m_tile_width);
-    uint64_t tile_start_offset = tile_size * tileIdx;
-
-
-    // --- read required file range
-
-    std::vector<uint8_t> src_data;
-    Error err = get_compressed_image_data_uncompressed(context, image_id, &src_data, tile_start_offset, tile_size, tileIdx, nullptr);
-    //Error err = context->get_heif_file()->append_data_from_iloc(image_id, src_data, tile_start_offset, tile_size);
-    if (err) {
-      return err;
-    }
-
-    UncompressedBitReader srcBits(src_data);
-
-    processTile(srcBits, tile_y, tile_x, out_x0, out_y0);
-
-    return Error::Ok;
-  }
-
-
-  void processTile(UncompressedBitReader& srcBits, uint32_t tile_row, uint32_t tile_column, uint32_t out_x0, uint32_t out_y0)
-  {
-    bool haveProcessedChromaForThisTile = false;
-    for (ChannelListEntry& entry : channelList) {
-      if (entry.use_channel) {
-        if ((entry.channel == heif_channel_Cb) || (entry.channel == heif_channel_Cr)) {
-          if (!haveProcessedChromaForThisTile) {
-            for (uint32_t tile_y = 0; tile_y < entry.tile_height; tile_y++) {
-              // TODO: row padding
-              uint64_t dst_row_number = tile_y + out_y0;
-              uint64_t dst_row_offset = dst_row_number * entry.dst_plane_stride;
-              for (uint32_t tile_x = 0; tile_x < entry.tile_width; tile_x++) {
-                uint64_t dst_column_number = out_x0 + tile_x;
-                uint64_t dst_column_offset = dst_column_number * entry.bytes_per_component_sample;
-                int val = srcBits.get_bits(entry.bytes_per_component_sample * 8);
-                memcpy(entry.dst_plane + dst_row_offset + dst_column_offset, &val, entry.bytes_per_component_sample);
-                val = srcBits.get_bits(entry.bytes_per_component_sample * 8);
-                memcpy(entry.other_chroma_dst_plane + dst_row_offset + dst_column_offset, &val, entry.bytes_per_component_sample);
-              }
-              haveProcessedChromaForThisTile = true;
-            }
-          }
-        }
-        else {
-          for (uint32_t tile_y = 0; tile_y < entry.tile_height; tile_y++) {
-            uint64_t dst_row_offset = entry.getDestinationRowOffset(tile_row, tile_y);
-            processComponentRow(entry, srcBits, dst_row_offset, tile_column);
-          }
-        }
-      }
-      else {
-        // skip over the data we are not using
-        srcBits.skip_bytes(entry.get_bytes_per_tile());
-        continue;
-      }
-    }
-  }
-};
 
 class RowInterleaveDecoder : public AbstractDecoder
 {
